Use integer gripper and DO targets

+++ lerobot_robot_romoya/lerobot_robot_romoya/robots/config_lebai_follower.py
@@ -15,14 +15,16 @@ class LebaiFollowerConfig(RobotConfig):
     acceleration: float = 1.0
     velocity: float = 1.0
     blend_radius: float = 0.0
-    gripper_force: float = 100.0
-    gripper_open_position: float = 99.0
-    gripper_closed_position: float = 0.0
+    gripper_force: int = 100
+    gripper_open_position: int = 99
+    gripper_closed_position: int = 0
     default_do0: int = 0
     default_do1: int = 0
     use_effort: bool = True
     use_velocity: bool = True
     use_acceleration: bool = True
+    # Deprecated: use top-level `--trace_path` from `lerobot-record` instead.
+    # Kept temporarily for compatibility so existing CLI invocations still parse.
     trace_path: Path | None = None
     cameras: dict[str, CameraConfig] = field(default_factory=dict)
